+++ tests/test_ProblemData.py
@@ -14,166 +14,6 @@ from pyvrp import (
     VehicleType,
 )
 
-_INT_MAX = np.iinfo(np.int64).max
-_MAX_SIZE = np.iinfo(np.uint64).max
-
-
-@pytest.mark.parametrize(
-    (
-        "location",
-        "delivery",
-        "pickup",
-        "service_duration",
-        "tw_early",
-        "tw_late",
-        "release_time",
-        "prize",
-        "required",
-        "group",
-        "name",
-    ),
-    [
-        (1, 1, 1, 1, 0, 1, 0, 0, True, None, "test name"),  # normal
-        (1, 1, 0, 0, 0, 1, 0, 0, True, None, "1234"),  # zero duration
-        (1, 0, 0, 1, 0, 1, 0, 0, True, None, "1,2,3,4"),  # zero delivery
-        (1, 1, 0, 1, 0, 0, 0, 0, True, None, ""),  # zero time windows
-        (1, 1, 0, 1, 0, 1, 1, 0, True, None, ""),  # positive release time
-        (0, 1, 0, 1, 0, 1, 0, 1, True, None, ""),  # positive prize
-        (0, 1, 0, 1, 0, 1, 0, 1, False, None, ""),  # not required
-        (0, 1, 0, 1, 0, 1, 0, 1, False, 0, ""),  # group membership
-    ],
-)
-def test_client_constructor_initialises_data_fields_correctly(
-    location: int,
-    delivery: int,
-    pickup: int,
-    service_duration: int,
-    tw_early: int,
-    tw_late: int,
-    release_time: int,
-    prize: int,
-    required: bool,
-    group: int | None,
-    name: str,
-):
-    """
-    Tests that the access properties return the data that was given to the
-    Client's constructor.
-    """
-    client = Client(
-        location=location,
-        delivery=[delivery],
-        pickup=[pickup],
-        service_duration=service_duration,
-        tw_early=tw_early,
-        tw_late=tw_late,
-        release_time=release_time,
-        prize=prize,
-        required=required,
-        group=group,
-        name=name,
-    )
-
-    assert_equal(client.location, location)
-    assert_equal(client.delivery, [delivery])
-    assert_equal(client.pickup, [pickup])
-    assert_equal(client.service_duration, service_duration)
-    assert_equal(client.tw_early, tw_early)
-    assert_equal(client.tw_late, tw_late)
-    assert_equal(client.release_time, release_time)
-    assert_equal(client.prize, prize)
-    assert_equal(client.required, required)
-    assert_equal(client.group, group)
-    assert_equal(client.name, name)
-    assert_equal(str(client), name)
-
-
-@pytest.mark.parametrize(
-    (
-        "delivery",
-        "pickup",
-        "service",
-        "tw_early",
-        "tw_late",
-        "release_time",
-        "prize",
-    ),
-    [
-        (1, 0, 0, 1, 0, 0, 0),  # late < early
-        (1, 0, 0, -1, 0, 0, 0),  # negative early
-        (0, 0, -1, 0, 1, 0, 1),  # negative service duration
-        (-1, 0, 1, 0, 1, 0, 0),  # negative delivery
-        (0, -1, 1, 0, 1, 0, 0),  # negative pickup
-        (0, 0, 0, 0, 1, -1, 0),  # negative release time
-        (0, 0, 0, 0, 1, 2, 0),  # release time > late
-        (1, 0, 1, 0, 1, 0, -1),  # negative prize
-    ],
-)
-def test_raises_for_invalid_client_data(
-    delivery: int,
-    pickup: int,
-    service: int,
-    tw_early: int,
-    tw_late: int,
-    release_time: int,
-    prize: int,
-):
-    """
-    Tests that invalid configurations are not accepted.
-    """
-    with assert_raises(ValueError):
-        Client(
-            0,
-            [delivery],
-            [pickup],
-            service,
-            tw_early,
-            tw_late,
-            release_time,
-            prize,
-        )
-
-
-@pytest.mark.parametrize(
-    ("tw_early", "tw_late", "service_duration"),
-    [
-        (1, 0, 0),  # tw_early > tw_late
-        (-1, 0, 0),  # tw_early < 0
-        (0, -1, 0),  # tw_late < 0
-        (0, 0, -1),  # service_duration < 0
-    ],
-)
-def test_raises_for_invalid_depot_data(
-    tw_early: int,
-    tw_late: int,
-    service_duration: int,
-):
-    """
-    Tests that an invalid depot configuration is not accepted.
-    """
-    with assert_raises(ValueError):
-        Depot(0, tw_early, tw_late, service_duration)
-
-
-def test_depot_initialises_data_correctly():
-    """
-    Tests that the depot constructor correctly initialises its member data, and
-    ensures the data is accessible from Python.
-    """
-    depot = Depot(
-        location=2,
-        tw_early=5,
-        tw_late=7,
-        service_duration=3,
-        name="test",
-    )
-
-    assert_equal(depot.location, 2)
-    assert_equal(depot.tw_early, 5)
-    assert_equal(depot.tw_late, 7)
-    assert_equal(depot.service_duration, 3)
-    assert_equal(depot.name, "test")
-
 
 def test_problem_data_raises_when_no_depot_is_provided():
     """
@@ -501,245 +341,6 @@ def test_matrices_are_not_copies():
     assert_(dur1.base is dur2.base)
 
 
-@pytest.mark.parametrize(
-    (
-        "capacity",
-        "num_available",
-        "tw_early",
-        "tw_late",
-        "shift_duration",
-        "max_distance",
-        "fixed_cost",
-        "unit_distance_cost",
-        "unit_duration_cost",
-        "start_late",
-        "initial_load",
-    ),
-    [
-        (0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0),  # num_available must be positive
-        (-1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0),  # capacity cannot be negative
-        (-100, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0),  # this is just wrong
-        (0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0),  # early > start late
-        (0, 1, 1, 1, 0, 0, 0, 0, 0, 2, 0),  # start late > late
-        (0, 1, -1, 0, 0, 0, 0, 0, 0, 0, 0),  # negative early
-        (0, 1, 0, -1, 0, 0, 0, 0, 0, 0, 0),  # negative late
-        (0, 1, 0, 0, -1, 0, 0, 0, 0, 0, 0),  # negative shift_duration
-        (0, 1, 0, 0, 0, -1, 0, 0, 0, 0, 0),  # negative max_distance
-        (0, 1, 0, 0, 0, 0, -1, 0, 0, 0, 0),  # negative fixed_cost
-        (0, 1, 0, 0, 0, 0, 0, -1, 0, 0, 0),  # negative unit_distance_cost
-        (0, 1, 0, 0, 0, 0, 0, 0, -1, 0, 0),  # negative unit_duration_cost
-        (0, 1, 0, 0, 0, 0, 0, 0, 0, -1, 0),  # negative start late
-        (0, 1, 0, 0, 0, 0, 0, 0, 0, 0, -1),  # negative initial load
-        (0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 2),  # initial load exceeds capacity
-    ],
-)
-def test_vehicle_type_raises_invalid_data(
-    capacity: int,
-    num_available: int,
-    tw_early: int,
-    tw_late: int,
-    shift_duration: int,
-    max_distance: int,
-    fixed_cost: int,
-    unit_distance_cost: int,
-    unit_duration_cost: int,
-    start_late: int,
-    initial_load: int,
-):
-    """
-    Tests that the vehicle type constructor raises when given invalid
-    arguments.
-    """
-    with assert_raises(ValueError):
-        VehicleType(
-            num_available=num_available,
-            capacity=[capacity],
-            fixed_cost=fixed_cost,
-            tw_early=tw_early,
-            tw_late=tw_late,
-            shift_duration=shift_duration,
-            max_distance=max_distance,
-            unit_distance_cost=unit_distance_cost,
-            unit_duration_cost=unit_duration_cost,
-            start_late=start_late,
-            initial_load=[initial_load],
-        )
-
-
-@pytest.mark.parametrize(
-    ("max_overtime", "unit_overtime_cost"),
-    [(-1, 0), (0, -1)],
-)
-def test_vehicle_type_raises_negative_overtime_data(
-    max_overtime: int,
-    unit_overtime_cost: int,
-):
-    with assert_raises(ValueError):
-        VehicleType(
-            max_overtime=max_overtime,
-            unit_overtime_cost=unit_overtime_cost,
-        )
-
-
-def test_vehicle_type_does_not_raise_for_all_zero_edge_case():
-    """
-    The vehicle type constructor should allow the following edge case where all
-    data has been zeroed out.
-    """
-    vehicle_type = VehicleType(
-        num_available=1,
-        capacity=[],
-        start_depot=0,
-        end_depot=0,
-        fixed_cost=0,
-        tw_early=0,
-        tw_late=0,
-        shift_duration=0,
-        max_distance=0,
-        unit_distance_cost=0,
-        unit_duration_cost=0,
-        start_late=0,
-    )
-
-    assert_equal(vehicle_type.num_available, 1)
-    assert_equal(vehicle_type.start_depot, 0)
-    assert_equal(vehicle_type.end_depot, 0)
-    assert_equal(vehicle_type.capacity, [])
-    assert_equal(vehicle_type.fixed_cost, 0)
-    assert_equal(vehicle_type.tw_early, 0)
-    assert_equal(vehicle_type.tw_late, 0)
-    assert_equal(vehicle_type.shift_duration, 0)
-    assert_equal(vehicle_type.max_distance, 0)
-    assert_equal(vehicle_type.unit_distance_cost, 0)
-    assert_equal(vehicle_type.unit_duration_cost, 0)
-    assert_equal(vehicle_type.start_late, 0)
-
-
-def test_vehicle_type_default_values():
-    """
-    Tests that the default values for costs and shift time windows are set
-    correctly.
-    """
-    vehicle_type = VehicleType()
-    assert_equal(vehicle_type.num_available, 1)
-    assert_equal(vehicle_type.start_depot, 0)
-    assert_equal(vehicle_type.end_depot, 0)
-    assert_equal(vehicle_type.capacity, [])
-    assert_equal(vehicle_type.fixed_cost, 0)
-    assert_equal(vehicle_type.tw_early, 0)
-    assert_equal(vehicle_type.unit_distance_cost, 1)
-    assert_equal(vehicle_type.unit_duration_cost, 0)
-    assert_equal(vehicle_type.unit_overtime_cost, 0)
-    assert_equal(vehicle_type.name, "")
-
-    # The default value for the following fields is the largest representable
-    # integral value.
-    assert_equal(vehicle_type.tw_late, _INT_MAX)
-    assert_equal(vehicle_type.shift_duration, _INT_MAX)
-    assert_equal(vehicle_type.max_duration, _INT_MAX)
-    assert_equal(vehicle_type.max_distance, _INT_MAX)
-
-    # The default value for start_late is the value of tw_late.
-    assert_equal(vehicle_type.start_late, vehicle_type.tw_late)
-
-
-def test_vehicle_type_attribute_access():
-    """
-    Smoke test that checks all attributes are equal to the values they were
-    given in the constructor's arguments.
-    """
-    vehicle_type = VehicleType(
-        num_available=7,
-        start_depot=29,
-        end_depot=43,
-        capacity=[13],
-        fixed_cost=3,
-        tw_early=17,
-        tw_late=19,
-        shift_duration=23,
-        max_distance=31,
-        unit_distance_cost=37,
-        unit_duration_cost=41,
-        start_late=18,
-        max_overtime=43,
-        name="vehicle_type name",
-    )
-
-    assert_equal(vehicle_type.num_available, 7)
-    assert_equal(vehicle_type.start_depot, 29)
-    assert_equal(vehicle_type.end_depot, 43)
-    assert_equal(vehicle_type.capacity, [13])
-    assert_equal(vehicle_type.fixed_cost, 3)
-    assert_equal(vehicle_type.tw_early, 17)
-    assert_equal(vehicle_type.tw_late, 19)
-    assert_equal(vehicle_type.shift_duration, 23)
-    assert_equal(vehicle_type.max_distance, 31)
-    assert_equal(vehicle_type.unit_distance_cost, 37)
-    assert_equal(vehicle_type.unit_duration_cost, 41)
-    assert_equal(vehicle_type.start_late, 18)
-    assert_equal(vehicle_type.max_overtime, 43)
-
-    assert_equal(vehicle_type.name, "vehicle_type name")
-    assert_equal(str(vehicle_type), "vehicle_type name")
-
-
-@pytest.mark.parametrize(
-    ("shift_duration", "max_overtime", "expected"),
-    [
-        (_INT_MAX, _INT_MAX, _INT_MAX),  # should not overflow
-        (_INT_MAX, 0, _INT_MAX),  # borderline
-        (0, _INT_MAX, _INT_MAX),  # borderline
-        (_INT_MAX - 1, 1, _INT_MAX),  # check for off-by-one
-        (1, _INT_MAX - 1, _INT_MAX),  # check for off-by-one
-        (10, 10, 20),  # completely OK, should sum both terms
-    ],
-)
-def test_vehicle_type_max_duration(
-    shift_duration: int,
-    max_overtime: int,
-    expected: int,
-):
-    """
-    Tests that the maximum duration property is correctly computed, and does
-    not over- or underflow.
-    """
-    veh_type = VehicleType(
-        shift_duration=shift_duration,
-        max_overtime=max_overtime,
-    )
-
-    assert_equal(veh_type.shift_duration, shift_duration)
-    assert_equal(veh_type.max_overtime, max_overtime)
-    assert_equal(veh_type.max_duration, expected)
-
-
-def test_vehicle_type_replace():
-    """
-    Tests that calling replace() on a VehicleType functions correctly.
-    """
-    vehicle_type = VehicleType(num_available=7, capacity=[10], name="test")
-    assert_equal(vehicle_type.num_available, 7)
-    assert_equal(vehicle_type.capacity, [10])
-    assert_equal(vehicle_type.name, "test")
-
-    # Replacing the number of available vehicles and name should be reflected
-    # in the returned vehicle type, but any other values should remain the same
-    # as the original. In particular, capacity should not be changed.
-    new = vehicle_type.replace(num_available=5, name="new")
-    assert_equal(new.num_available, 5)
-    assert_equal(new.capacity, [10])
-    assert_equal(new.name, "new")
-
-
-def test_vehicle_type_multiple_capacities():
-    """
-    Tests that vehicle types correctly handle multiple capacities.
-    """
-    vehicle_type = VehicleType(capacity=[998, 37], num_available=10)
-    assert_equal(vehicle_type.num_available, 10)
-    assert_equal(vehicle_type.capacity, [998, 37])
-
-
 def test_depot_client_raises_invalid_index(ok_small):
     """
     Tests that calling depot(idx) and client(idx) raises when the index is out
@@ -820,19 +421,6 @@ def test_raises_inconsistent_profiles(ok_small):
     """
     with assert_raises(ValueError):
         ok_small.replace(distance_matrices=ok_small.distance_matrices() * 2)
-
-
-def test_client_group_attribute_access():
-    """
-    Tests that the ClientGroup's attributes are correctly set and accessible.
-    """
-    group = ClientGroup(clients=[1, 2, 3], required=False, name="test")
-
-    assert_equal(group.clients, [1, 2, 3])
-    assert_equal(group.required, False)
-    assert_equal(group.mutually_exclusive, True)
-    assert_equal(group.name, "test")
-    assert_equal(str(group), "test")
 
 
 def test_raises_empty_group():
@@ -950,20 +538,6 @@ def test_raises_for_required_mutually_exclusive_group_membership():
         )
 
 
-def test_client_group_raises_duplicate_clients():
-    """
-    Tests that adding the same client to a group more than once raises.
-    """
-    with assert_raises(ValueError):
-        ClientGroup([1, 1])
-
-    group = ClientGroup()
-    group.add_client(1)  # this should be OK
-
-    with assert_raises(ValueError):
-        group.add_client(1)  # but adding the client a second time is not
-
-
 def test_replacing_client_groups(ok_small):
     """
     Tests that replacing mutually exclusive client groups works well.
@@ -983,121 +557,6 @@ def test_replacing_client_groups(ok_small):
     assert_equal(data.group(0).clients, [0])
 
 
-def test_client_eq():
-    """
-    Tests the client's equality operator.
-    """
-    client1 = Client(location=0, delivery=[1], pickup=[2], tw_late=3, group=0)
-    client2 = Client(location=0, delivery=[1], pickup=[2], tw_late=3, group=1)
-    assert_(client1 != client2)
-
-    # This client is equivalent to client1.
-    client3 = Client(location=0, delivery=[1], pickup=[2], tw_late=3, group=0)
-    assert_(client1 == client3)
-    assert_(client3 == client3)
-
-    # And some things that are not clients.
-    assert_(client1 != "text")
-    assert_(client1 != 1)
-
-
-def test_depot_eq():
-    """
-    Tests the depot's equality operator.
-    """
-    depot1 = Depot(location=0)
-    depot2 = Depot(location=1)
-    assert_(depot1 != depot2)
-
-    # This depot is equivalent to depot1.
-    depot3 = Depot(location=0)
-    assert_(depot1 == depot3)
-    assert_(depot3 == depot3)
-
-    # And some things that are not depots.
-    assert_(depot1 != "text")
-    assert_(depot1 != 3)
-
-    depot4 = Depot(location=0, name="test")
-    assert_(depot1 != depot4)
-
-
-def test_vehicle_type_eq():
-    """
-    Tests the vehicle type's equality operator.
-    """
-    veh_type1 = VehicleType(num_available=3, profile=0)
-    veh_type2 = VehicleType(num_available=3, profile=1)
-    assert_(veh_type1 != veh_type2)
-
-    # This vehicle type is equivalent to veh_type1.
-    veh_type3 = VehicleType(num_available=3, profile=0)
-    assert_(veh_type1 == veh_type3)
-
-    # And some things that are not vehicle types.
-    assert_(veh_type1 != "text")
-    assert_(veh_type1 != 5)
-
-
-def test_client_group_eq():
-    """
-    Tests the client group's equality operator.
-    """
-    group1 = ClientGroup(clients=[1, 2], required=False)
-    group2 = ClientGroup(clients=[1, 2], required=True)
-    assert_(group1 != group2)
-
-    # This group is equivalent to group1.
-    group3 = ClientGroup(clients=[1, 2], required=False)
-    assert_(group1 == group3)
-
-    # And some things that are not client groups.
-    assert_(group1 != "text")
-    assert_(group1 != 7)
-
-    group4 = ClientGroup(clients=[1, 2], required=False, name="test")
-    assert_(group1 != group4)
-
-
-def test_eq_checks_names():
-    """
-    Tests that the equality operators on named objects also considers the name
-    when determining equality.
-    """
-    assert_(Client(0, name="1") != Client(0, name="2"))
-    assert_(Depot(0, name="1") != Depot(0, name="2"))
-    assert_(VehicleType(name="1") != VehicleType(name="2"))
-    assert_(ClientGroup(name="1") != ClientGroup(name="2"))
-
-
-@pytest.mark.parametrize("cls", (Client, Depot))
-def test_pickle_locations(cls):
-    """
-    Tests that client and depot locations can be serialised and unserialised.
-    """
-    before_pickle = cls(location=0, name="test")
-    bytes = pickle.dumps(before_pickle)
-    assert_equal(pickle.loads(bytes), before_pickle)
-
-
-def test_pickle_client_group():
-    """
-    Tests that client groups can be serialised and unserialised.
-    """
-    before_pickle = ClientGroup(clients=[1, 2, 3], required=False, name="test")
-    bytes = pickle.dumps(before_pickle)
-    assert_equal(pickle.loads(bytes), before_pickle)
-
-
-def test_pickle_vehicle_type():
-    """
-    Tests that vehicle types can be serialised and unserialised.
-    """
-    before_pickle = VehicleType(num_available=12, capacity=[3], name="test123")
-    bytes = pickle.dumps(before_pickle)
-    assert_equal(pickle.loads(bytes), before_pickle)
-
-
 def test_pickle_data(ok_small, rc208):
     """
     Tests that problem data instances can be serialised and unserialised.
@@ -1107,56 +566,6 @@ def test_pickle_data(ok_small, rc208):
 
     bytes = pickle.dumps(rc208)
     assert_equal(pickle.loads(bytes), rc208)
-
-
-@pytest.mark.parametrize(
-    ("delivery", "pickup", "exp_delivery", "exp_pickup"),
-    [
-        ([0], [0], [0], [0]),
-        ([0], [0, 1, 2], [0, 0, 0], [0, 1, 2]),
-        ([0, 1, 2], [0], [0, 1, 2], [0, 0, 0]),
-        ([0, 2], [1], [0, 2], [1, 0]),
-        ([], [], [], []),
-    ],
-)
-def test_client_load_dimensions_are_padded_with_zeroes(
-    delivery: list[int],
-    pickup: list[int],
-    exp_delivery: list[int],
-    exp_pickup: list[int],
-):
-    """
-    Tests that any missing load dimensions for the pickup and delivery Client
-    arguments are padded with zeroes.
-    """
-    client = Client(0, delivery=delivery, pickup=pickup)
-    assert_equal(client.delivery, exp_delivery)
-    assert_equal(client.pickup, exp_pickup)
-
-
-@pytest.mark.parametrize(
-    ("capacity", "initial_load", "exp_capacity", "exp_initial_load"),
-    [
-        ([0], [0], [0], [0]),
-        ([0], [0, 0, 0], [0, 0, 0], [0, 0, 0]),
-        ([0, 1, 2], [0], [0, 1, 2], [0, 0, 0]),
-        ([1, 2], [1], [1, 2], [1, 0]),
-        ([], [], [], []),
-    ],
-)
-def test_vehicle_load_dimensions_are_padded_with_zeroes(
-    capacity: list[int],
-    initial_load: list[int],
-    exp_capacity: list[int],
-    exp_initial_load: list[int],
-):
-    """
-    Tests that any missing load dimensions for the capacity and initial_load
-    VehicleType arguments are padded with zeroes.
-    """
-    vehicle_type = VehicleType(capacity=capacity, initial_load=initial_load)
-    assert_equal(vehicle_type.capacity, exp_capacity)
-    assert_equal(vehicle_type.initial_load, exp_initial_load)
 
 
 def test_problem_data_raises_when_pickup_and_delivery_dimensions_differ():
@@ -1301,34 +710,6 @@ def test_validate_raises_for_invalid_reload_depot(ok_small):
     # during argument validation.
     with assert_raises(IndexError):
         ok_small.replace(vehicle_types=[new_vehicle_type])
-
-
-def test_vehicle_type_max_trips(ok_small_multiple_trips):
-    """
-    Tests that the vehicle type correctly handles the case where max_reloads
-    is set to its largest allowed size - then max_trips should not overflow.
-    """
-    veh_type = ok_small_multiple_trips.vehicle_type(0)
-    assert_equal(veh_type.max_reloads, 1)
-    assert_equal(veh_type.max_trips, 2)
-
-    # Normally, max_trips == max_reloads + 1, but when max_reloads is at the
-    # maximum size, we do not want max_trips to overflow and wrap around to
-    # zero. These asserts check that does not happen.
-    veh_type = veh_type.replace(max_reloads=_MAX_SIZE)
-    assert_equal(veh_type.max_reloads, _MAX_SIZE)
-    assert_equal(veh_type.max_trips, _MAX_SIZE)
-
-
-def test_vehicle_max_trips_is_one_if_no_reload_depots(ok_small):
-    """
-    Tests that a vehicle type's max_trips is one if there's no reload depots,
-    despite max_reloads being unconstrained.
-    """
-    veh_type = ok_small.vehicle_type(0)
-    assert_equal(veh_type.reload_depots, [])
-    assert_equal(veh_type.max_reloads, _MAX_SIZE)
-    assert_equal(veh_type.max_trips, 1)
 
 
 def test_has_time_windows(small_cvrp, pr107, small_spd, ok_small, rc208):
